gsm_instance::proton

Function list_available_releases

pub fn list_available_releases() -> Result<Vec<ProtonRelease>, ReleaseError>

Fetches a list of all available Proton GE releases from GitHub.

This function makes a request to the GitHub API to get a list of all releases for the proton-ge-custom repository. It then filters these releases to find the ones that have a .tar.gz asset, which is the format used for distribution.